+++ services/horizon/internal/db2/schema/migrations/35_drop_participant_id.sql
@@ -1,0 +1,15 @@
+-- +migrate Up
+
+ALTER TABLE history_operation_participants
+    DROP COLUMN id;
+
+ALTER TABLE history_transaction_participants
+    DROP COLUMN id;
+
+-- +migrate Down
+
+ALTER TABLE history_operation_participants
+    ADD COLUMN id integer;
+
+ALTER TABLE history_transaction_participants
+    ADD COLUMN id integer;